requirements.txt
apache-airflow==2.6.3
requests==2.31.0
pandas==2.0.3
python-dotenv==1.0.0
plotly==5.15.0
dash==2.11.1